Commentaire, instanciation de la table de level

+++ Anima_MJ/Donnees.cs
@@ -281,7 +281,7 @@ namespace Anima_MJ
         public class Perso
         {
             
-            private long id = 1;
+            private int id = 1;
             private string joueur;
             private string nom;
             private string prnom;
@@ -304,11 +304,7 @@ namespace Anima_MJ
             private int apparence;
             private int xp;
             private int pf;
-            //private int di;
-            //private int ppp;
-            //private int charac;
-            //private int characNaturel;
-            //private int compNaturel;
+           
 
             public string Joueur { get => joueur; set => joueur = value; }
             public string Nom { get => nom; set => nom = value; }
@@ -331,7 +327,7 @@ namespace Anima_MJ
             public int Apparence { get => apparence; set => apparence = value; }
             public int Xp { get => xp; set => xp = value; }
             public int Hauteur { get => hauteur; set => hauteur = value; }
-            public long Id { get => id; set => id = value; }
+            public int Id { get => id; set => id = value; }
             public int Pf { get => pf; set => pf = value; }
         }
     }
